WordBase v2
Java file for the re-hauled version of my first design. 

Instructions are such: in the input file, on one line (doesn't matter what line you start from), put the name of what your save file to be and the first letter of what team you are. (ex: "muffie O"). On the line directly below that, input the letters in your game board, spaces between letters. After that, skip one line, and input the corresponding points for the type of tile: 0 if you own it, 1 if it's white, 2 if the opponent owns it, and 3 if it's a bomb.

Features:
- load game system so you can have as many games going as you want at a time 
- finds words with the greatest overall length (letter count)
- finds words that get the closest to the opponent's base
- option to run a specific tile search that looks for words that land on or go through a certain tile on the board, useful for cutting off the opponent
- option to change the list size to give you more options (top 10, top 15, top 50, etc)

To-Do List:
- add a new criteria: tiles gained. this includes tiles gained by bombs and discounts tiles that you already owned.
- add a new criteria: opponent tiles lost. this includes cut-off letters.
- fantastical ideas: create a scoring system that weights all these different criteria together to generate the "best word". expand to program to pit AI's with different weighted scoring systems against each other to determine which scoring system is truly the best.